Landing Page and Navigation Page
I took today to try out a rolling title feature that transitions into a
navigation page for my portfolio. Messing around with the timing proved
difficult on the landing page and I came to the conclusion that instead
of trying to calculate specific ms I should account for certain windows
in which the ms may have an error. These windows are hardcoded values
that make it so that the animations will run properly